requirements.txt
openai
azure-cognitiveservices-speech
spidev
RPi.GPIO